clang compatibility

+++ Code/CryCommon/CryRenderer/IRenderAuxGeom.h
@@ -281,7 +281,7 @@ SAuxGeomRenderFlags::GetMode2D3DFlag() const
 	uint32 mode2D3D( m_renderFlags & e_Mode2D3DMask );
 	switch( mode2D3D )
 	{
-	case e_Mode2D:
+	case (uint32_t)e_Mode2D:
 		{
 			return( e_Mode2D );
 		}
